+++ Kumbajah.Domain/Entities/Product.cs
@@ -8,7 +8,7 @@
         public string Image { get; private set; }
         public long Quantity { get; private set; }
         public virtual Category Category { get; private set; }
-        public int CategoryId { get; private set; }
+        public long CategoryId { get; private set; }
 
         public Product() { }
